refactor<config>: inline default struct
